+++ tests/testthat/test_project.R
@@ -41,13 +41,13 @@ if (require(rstanarm)) {
   vs_list <- lapply(fit_list, varsel, nv_max = nv, verbose = FALSE)
   
   
-  test_that("project: relaxing has the expected effect", {
+  test_that("project: cv_searching has the expected effect", {
     
     vs_list <- lapply(fit_list, varsel, nv_max = nv, verbose = FALSE, method='l1')
     for (i in seq_along(vs_list)) {
       
-      p0 <- project(vs_list[[i]], relax=F, nv=1:nv)
-      p1 <- project(vs_list[[i]], relax=T, nv=1:nv, nc=1, regul=1e-9)
+      p0 <- project(vs_list[[i]], cv_search=F, nv=1:nv)
+      p1 <- project(vs_list[[i]], cv_search=T, nv=1:nv, nc=1, regul=1e-9)
       
       for (j in seq_along(p1)) {
         # L1-penalised coefficients should have smaller L1-norm